# create-a-container/openapi.v1.yaml
openapi: 3.0.3
info:
  title: Create-a-Container API v1
  description: |
    JSON-only REST API for the React SPA and external automation.
    Co-exists with the legacy `/` (HTML) and `/apikeys` (mixed) routes during the React migration.

    ## Authentication

    Two mechanisms are supported:

    - **Session cookie** — established by `POST /api/v1/auth/login`. Requires a CSRF token (see below) on every state-changing request.
    - **API key** — `Authorization: Bearer <key>`. CSRF is skipped for Bearer-authenticated requests.

    ## CSRF (session auth only)

    1. `GET /api/v1/csrf-token` → returns `{ "data": { "csrfToken": "..." } }` and sets `__Host-csrf.token` cookie.
    2. Send the token in the `X-CSRF-Token` header on every `POST`/`PUT`/`PATCH`/`DELETE`.

    ## Response envelope

    Success: `{ "data": <payload>, "meta"?: {...} }`.
    Error:   `{ "error": { "code": "...", "message": "...", "fields"?: {...} } }`.
  version: 1.0.0
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    SessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
...
security:
  - BearerAuth: []
  - SessionCookie: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Username/password login (may require 2FA)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
                redirect: { type: string }
      responses:
        '200': { description: Logged in OR 2FA challenge issued }
        '401': { description: Invalid credentials }
  /auth/login/challenge/{id}:
    get:
      tags: [Auth]
      summary: Poll 2FA challenge status
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: status -> pending|approved|rejected|timeout|failed|unregistered }
...
  /auth/register:
    post:
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [uid, givenName, sn, mail, userPassword]
              properties:
                uid: { type: string }
                givenName: { type: string }
                sn: { type: string }
                mail: { type: string, format: email }
                userPassword: { type: string, format: password }
                inviteToken: { type: string }
      responses:
        '201': { description: Account created }
  /auth/register/invite/{token}:
    get:
      tags: [Auth]
      security: []
      parameters: [{ in: path, name: token, required: true, schema: { type: string } }]
      responses: { '200': { description: Invite valid, returns email } }
  /auth/register/2fa-qr/{token}:
    get:
      tags: [Auth]
      security: []
      parameters: [{ in: path, name: token, required: true, schema: { type: string } }]
      responses: { '200': { description: QR code data URI for push-notification enrollment } }
  /auth/password-reset/request:
    post:
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [usernameOrEmail]
              properties:
                usernameOrEmail: { type: string }
      responses: { '200': { description: Generic OK (does not reveal account existence) } }
  /auth/password-reset/{token}:
    get:
      tags: [Auth]
      security: []
      parameters: [{ in: path, name: token, required: true, schema: { type: string } }]
      responses:
        '200': { description: Token valid }
        '404': { description: Invalid or expired }
    post:
      tags: [Auth]
      security: []
      parameters: [{ in: path, name: token, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password, confirmPassword]
              properties:
                password: { type: string }
                confirmPassword: { type: string }
      responses: { '200': { description: Password reset } }
...